Fixed existing velocity after dying bug. Added testing mode
Player now dies and does not get respawned when checkpoint manager is turned off. A public bool has been added to the car (Testing Mode) that when turned on prevents the player from dying when checkpoint manager is off, meaning you can still move the car freely around the level.